[DataGrid] Move virtualization logic to hook
