+++ packages/grid/_modules_/grid/components/base/GridBody.tsx
@@ -9,7 +9,6 @@ import { GridMainContainer } from '../containers/GridMainContainer';
 import { GridAutoSizer } from '../GridAutoSizer';
 import { GridOverlays } from './GridOverlays';
 import { useGridRootProps } from '../../hooks/utils/useGridRootProps';
-import { GridVirtualScroller } from '../GridVirtualScroller';
 import { useGridSelector } from '../../hooks/utils/useGridSelector';
 import { gridSelectionStateSelector } from '../../hooks/features/selection/gridSelectionSelector';
 import { gridDensityHeaderHeightSelector } from '../../hooks/features/density/densitySelector';
@@ -17,10 +16,17 @@ import { GridScrollArea } from '../GridScrollArea';
 
 interface GridBodyProps {
   children?: React.ReactNode;
+  VirtualScrollerComponent: React.JSXElementConstructor<
+    React.HTMLAttributes<HTMLDivElement> & {
+      ref: React.Ref<HTMLDivElement>;
+      selectionLookup: {};
+      disableVirtualization: boolean;
+    }
+  >;
 }
 
 function GridBody(props: GridBodyProps) {
-  const { children } = props;
+  const { children, VirtualScrollerComponent } = props;
   const apiRef = useGridApiContext();
   const rootProps = useGridRootProps();
   const selection = useGridSelector(apiRef, gridSelectionStateSelector);
@@ -98,10 +104,10 @@ function GridBody(props: GridBodyProps) {
             // In this case, let the container to grow whatever it needs.
             height: size.height ? size.height - headerHeight : 'auto',
             marginTop: headerHeight,
-          };
+          } as React.CSSProperties;
 
           return (
-            <GridVirtualScroller
+            <VirtualScrollerComponent
               ref={windowRef}
               style={style}
               selectionLookup={selectionLookup} // TODO pass it directly to the row via componentsProps
@@ -121,6 +127,7 @@ GridBody.propTypes = {
   // | To update them edit the TypeScript types and run "yarn proptypes"  |
   // ----------------------------------------------------------------------
   children: PropTypes.node,
+  VirtualScrollerComponent: PropTypes.elementType.isRequired,
 } as any;
 
 export { GridBody };
