+++ packages/grid/_modules_/grid/components/virtualization/GridVirtualScrollerContent.tsx
@@ -1,0 +1,44 @@
+import * as React from 'react';
+import clsx from 'clsx';
+import { styled } from '@mui/material/styles';
+import { unstable_composeClasses as composeClasses } from '@mui/material';
+import { useGridRootProps } from '../../hooks/utils/useGridRootProps';
+import { getDataGridUtilityClass } from '../../gridClasses';
+import { GridComponentProps } from '../../GridComponentProps';
+
+type OwnerState = { classes: GridComponentProps['classes'] };
+
+const useUtilityClasses = (ownerState: OwnerState) => {
+  const { classes } = ownerState;
+
+  const slots = {
+    root: ['virtualScrollerContent'],
+  };
+
+  return composeClasses(slots, getDataGridUtilityClass, classes);
+};
+
+const VirtualScrollerContentRoot = styled('div', {
+  name: 'MuiDataGrid',
+  slot: 'VirtualScrollerContent',
+  overridesResolver: (props, styles) => styles.virtualScrollerContent,
+})({
+  position: 'relative',
+  overflow: 'hidden',
+});
+
+const GridVirtualScrollerContent = React.forwardRef<
+  HTMLDivElement,
+  React.HTMLAttributes<HTMLDivElement>
+>(function GridVirtualScrollerContent(props, ref) {
+  const { className, ...other } = props;
+  const rootProps = useGridRootProps();
+  const ownerState = { classes: rootProps.classes };
+  const classes = useUtilityClasses(ownerState);
+
+  return (
+    <VirtualScrollerContentRoot ref={ref} className={clsx(classes.root, className)} {...other} />
+  );
+});
+
+export { GridVirtualScrollerContent };
